Cleanup Engset source.

+++ src/types.h
@@ -3,8 +3,8 @@
 
 #include <cstdint>
 
-using Time = double;
-using Duration = double;
+using Time = long double;
+using Duration = long double;
 using Size = int64_t;
 using Intensity = double;
 using Uuid = uint64_t;
